refactoring node to remove duplicate code and to make it nice
